chore: reuse config
