+++ packages/rspack-diff/fixtures/three/rspack.config.js
@@ -1,18 +1,1 @@
-/**
- * @type {import('webpack').Configuration}
- */
-module.exports = {
-	mode: "development",
-	devtool: false,
-	entry: {
-		main: "./index.js"
-	},
-	target: ["web", "es5"],
-	optimization: {
-		minimize: false,
-		concatenateModules: false
-	},
-	output: {
-		path: "rspack-dist"
-	}
-};
+module.exports = require("../rspack.config");
