+++ packages/rspack-diff/package.json
@@ -14,7 +14,8 @@
     "diff:antd": "node . dist fixtures/antd/rspack-dist/main.js fixtures/antd/webpack-dist/main.js",
     "diff:three": "node . dist fixtures/three/rspack-dist/main.js fixtures/three/webpack-dist/main.js",
     "diff:lodash": "node . dist fixtures/lodash/rspack-dist/main.js fixtures/lodash/webpack-dist/main.js",
-    "diff:lodash_stats": "node . stats fixtures/lodash/rspack-dist/stats.json fixtures/lodash/webpack-dist/stats.json"
+    "diff:lodash_stats": "node . stats fixtures/lodash/rspack-dist/stats.json fixtures/lodash/webpack-dist/stats.json",
+    "diff:antd_stats": "node . stats fixtures/antd/rspack-dist/stats.json fixtures/antd/webpack-dist/stats.json"
   },
   "files": [
     "src"
